app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy      # The tool that translates Python code to SQL
psycopg2-binary # The driver that allows Python to talk to PostgreSQL
celery
boto3
python-multipart
openai-whisper
transformers
torch